ADD: fix product update
