+++ src/components/partials/EditProduct.tsx
@@ -3,6 +3,7 @@ import { useAppSelector, useAppDispatch  } from '../../app/hooks';
 import {AiOutlineSave, AiFillCloseCircle} from 'react-icons/ai';
 import Product from '../../Model/Product';
 import { editProduct } from '../../features/products/productSlice';
+import { updateProduct} from '../../api/products';
 
 type editProductProps =  { 
     product: Product, 
@@ -14,28 +15,41 @@ const EditProduct = ({product, setIsEditable}: editProductProps) => {
 
     const {id, name, image, price, category, stock } = product;
     const categories = useAppSelector((state) => state.products.categories);
-    const [productName, setProductName] = useState<string>(name);
-    const [productPrice, setProductPrice] = useState<Number>(price);
-    const [stockQuantity, setStockQuantity] = useState<Number>(stock);
+
+    const [editedName, setEditedName] = useState<string| ''>(name);
+    // const [editedImage, setEditedImage] = useState<string| ''>(image);
+    const [editedPrice, setEditedPrice] = useState<Number| 0>(price);
+    // const [editedCategory, setEditedCategory] = useState<string| ''>(category);
+    const [editedStock, setEditedStock] = useState<Number| ''>(stock);
 
     const handleCloseModal = () => setIsEditable(false);
-    const handleSaveEdits = () => {
-        const updateProduct = {
-            id,
-            "name": productName,
-            "image": image, //update this to be able to update image
-            "price": productPrice,
-            category, // this to will need to be updare
-            "stock": stockQuantity
+
+    const handleSaveEdits = async () => {
+        const editedProduct = {
+            "id": id,
+            "name": editedName,
+            "image": image, //editedImage,
+            "price": +editedPrice,
+            "category": category, // editedCategory,
+            "stock": +editedStock
         }
-        dispatch(editProduct(updateProduct));
+
+        // console.log(editProduct); 
+
+        const result =  await updateProduct(editedProduct);
+        if(!result) return "Product update failed"
+        //if product is update on the server, update the product on the client
+        dispatch(editProduct(result));
         setIsEditable(false);
     }
+    
 
     // handle form field updates
-    const handleProductNameUpdate = (event: ChangeEvent<HTMLInputElement>) => setProductName(event.target.value);
-    const handlePriceUpdate = (event: ChangeEvent<HTMLInputElement>) => setProductPrice(+event.target.value);
-    const handleStockUpdate = (event: ChangeEvent<HTMLInputElement>) => setStockQuantity(+event.target.value);
+    const handleProductNameUpdate = (event: ChangeEvent<HTMLInputElement>) => setEditedName(event.target.value);
+    // const handleImageUpdate = (event: ChangeEvent<HTMLInputElement>) => setEditedImage(event.target.value);
+    const handlePriceUpdate = (event: ChangeEvent<HTMLInputElement>) => setEditedPrice(+event.target.value);
+    // const handleCategoryUpdate = (event: ChangeEvent<HTMLSelectElement>) => setEditedCategory(event.target.value);
+    const handleStockUpdate = (event: ChangeEvent<HTMLInputElement>) => setEditedStock(+event.target.value);
 
     return (
         <div className="overlay">
@@ -51,7 +65,7 @@ const EditProduct = ({product, setIsEditable}: editProductProps) => {
                 </label>
 
                 <label htmlFor="name">Product Name
-                    <input type="text" value={productName} name="name" id="name" onChange={(e) => handleProductNameUpdate(e)}/>
+                    <input type="text" value={editedName} name="name" id="name" onChange={(e) => handleProductNameUpdate(e)}/>
                 </label>
 
                 <div>
@@ -59,7 +73,7 @@ const EditProduct = ({product, setIsEditable}: editProductProps) => {
                 </div>
 
                 <label htmlFor="price">Price
-                    <input type="text" value={productPrice.toString()}  id="price" onChange={(e) => handlePriceUpdate(e)}/>
+                    <input type="text" value={editedPrice.toString()}  id="price" onChange={(e) => handlePriceUpdate(e)}/>
                 </label>
 
                 <label htmlFor="catgeory">Category
@@ -70,7 +84,7 @@ const EditProduct = ({product, setIsEditable}: editProductProps) => {
                 </label>
 
                 <label htmlFor="stock">Quantity in stock
-                    <input type="text" value={stockQuantity.toString()} onChange={(e) => handleStockUpdate(e)} />
+                    <input type="text" value={editedStock.toString()} onChange={(e) => handleStockUpdate(e)} />
                 </label>
 
                 <AiOutlineSave size="2em" onClick={() => handleSaveEdits()}/>
